MISC/Prescription_OCR/backend/requirements.txt
fastapi
uvicorn
opencv-python
pillow
numpy
layoutparser
openai
python-multipart